docs-requirements.txt
sphinx~=2.4
sphinx-rtd-theme~=0.4
sphinx-autodoc-typehints~=1.10.2
# External
opentelemetry-api>= 0.13b0
opentelemetry-sdk>= 0.13b0
psutil>= 5.6.3

